\name{MatrixRider-package}
\alias{MatrixRider-package}
\alias{MatrixRider}
\docType{package}
\title{
What the package does (short line)
TODO
}
\description{
TODO
}
\details{
TODO
}
\author{
Elena Grassi <elena.grassi@unito.it>
}
\references{
TODO
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation %directory ~~
\keyword{MatrixRider}
\seealso{
TODO
}
\examples{
#TODO
}
